% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_wallace.R
\name{adj_wallace}
\alias{adj_wallace}
\title{Calculate the adjusted Wallace coefficient}
\usage{
adj_wallace(clusters_A, clusters_B)
}
\arguments{
\item{clusters_A, clusters_B}{A set of cluster identities (cluster memberships) for each member in a
community.}
}
\value{
The adjusted Wallace coefficient.
}
\description{
Calculate the adjusted Wallace coefficient.
}
\details{
From dorbarker's nAWC \url{https://github.com/dorbarker/nAWC}.
}
