% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim_to_count_table_ls.R
\name{cim_to_count_table_ls}
\alias{cim_to_count_table_ls}
\title{Create a list of tables of counts by group and cluster}
\usage{
cim_to_count_table_ls(
  .cim,
  .group = group,
  .cols = tidyselect::starts_with("schema")
)
}
\arguments{
\item{.cim}{table: a community identity matrix (CIM) with a group column (.group), and one
or more schema columns (.cols).}

\item{.group}{unquoted expression: the group column name.}

\item{.cols}{tidyeval expression: the schema column name(s), in tidyeval notation.}
}
\value{
a list of tibbles (by schema), of counts of objects in each cluster by group.
}
\description{
Create a list (indexed by schema) of tables of counts of cluster-group
  pairs, from a community identity matrix.
}
