% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_to_cdm.R
\name{count_table_to_cdm}
\alias{count_table_to_cdm}
\title{Create a community data matrix (CIM) from a table of cluster-group counts}
\usage{
count_table_to_cdm(.counts, .clstr_col = clstr, .count_col = n)
}
\arguments{
\item{.counts}{dataframe: a table of counts by group.}

\item{.clstr_col}{unquoted expression: the column with cluster identities.}

\item{.count_col}{unquoted expression: the column with the counts.}
}
\value{
a community data matrix (CDM).
}
\description{
Create a community data matrix (CDM) from a table of counts by cluster and group.
}
