% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fns.R
\name{diversityindex}
\alias{diversityindex}
\alias{shannon}
\alias{simpsons}
\title{Calculate a diversity index}
\usage{
shannon(clusters, base = 2)

simpsons(clusters)
}
\arguments{
\item{clusters}{A set of cluster identities (cluster memberships) for each member in a
community.}

\item{base}{The log-base.}
}
\value{
The diversity index.
}
\description{
Calculate a diversity index for a community.
}
\details{
From dorbarker's nAWC \url{https://github.com/dorbarker/nAWC}.
}
