% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_compare_forms.R
\name{make_plot_compare_forms}
\alias{make_plot_compare_forms}
\title{Plot a comparison between beta measures for abundance data and their presence-absence forms}
\usage{
make_plot_compare_forms(fndata, .measure, .comparison)
}
\arguments{
\item{fndata}{A table of measures of beta diversity in long format, with columns: `schema_order`,
`measure`, `measure_type`, `value`, `comparison`.}

\item{.measure}{A vector of the name(s) of measure(s) to plot. If length > 1, plot will be
faceted by .measure. If .measure is not specified, facet across all measures
in fndata.}

\item{.comparison}{A vector of the name(s) of comparison(s) to plot. If length > 1, plot will be
faceted by .comparison If .comparison is not specified, facet across all
comparison in fndata.}
}
\value{
A step plot of diversity.
}
\description{
Compare measures of beta diversity for abundance (count) measures and their
   presence-absence (binary) forms.
}
\details{
fndata must be in long format, with columns:  `schema_order`,
   `measure`, `measure_type`, `value`, `comparison`.
}
