% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_diversity_var.R
\name{make_plot_diversity_var}
\alias{make_plot_diversity_var}
\title{Plot the variability in measures between comparisons}
\usage{
make_plot_diversity_var(fndata)
}
\arguments{
\item{fndata}{A table of measures of beta diversity in long format, with columns: `schema_order`,
`measure`, `measure_type`, `value`, `comparison`.}
}
\value{
A box plot of variability by schema.
}
\description{
Plot the variability in measures between comparisons (between groups).
}
\details{
fndata must be in long format, with columns:  `schema_order`,
   `measure`, `measure_type`, `value`, `comparison`.
}
