% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_heat.R
\name{make_plot_heat}
\alias{make_plot_heat}
\title{Plot measures of beta diversity for multiple comparisons as a heat map}
\usage{
make_plot_heat(fndata, .measure, .schema, .lblsize = 5)
}
\arguments{
\item{fndata}{A table of measures of beta diversity in long format, with columns: `schema`,
`measure`, `value`, `group_from`, `group_to`.}

\item{.measure}{A vector of the name(s) of measure(s) to plot. If length > 1, plot will be
faceted by .measure. If .measure is not specified, facet across all measures
in fndata.}

\item{.schema}{A vector of the name(s) of schema(s) to plot. If length > 1, plot will be
faceted by .schema If .schema is not specified, facet across all schema
in fndata.}

\item{.lblsize}{A label size (numeric) for numeric labels.}
}
\value{
A heat map of diversity between stages.
}
\description{
make_plot_heat() creates a heat map of diversity between from a long
   data frame of diversity measures.
}
\details{
fndata must be in long format, with columns:  'schema', 'measure',
   'value', 'group_from`, `group_to`.

   This visualization is useful where contiguous comparisons are on the
   diagonal. They will be on the diagonal if they are adjacent in order of
   appearance in 'group_from'.
}
