% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_net.R
\name{make_plot_net}
\alias{make_plot_net}
\title{Plot measures of beta diversity for multiple comparisons as a network diagram}
\usage{
make_plot_net(fndata, .measure, .schema, .lnsize = c(1, 5))
}
\arguments{
\item{fndata}{A table of measures of beta diversity in long format, with columns: `schema`,
`measure`, `value`, `group_from`, `group_to`.}

\item{.measure}{A vector of the name(s) of measure(s) to plot. If length > 1, plot will be
faceted by .measure. If .measure is not specified, facet across all measures
in fndata.}

\item{.schema}{A vector of the name(s) of schema(s) to plot. If length > 1, plot will be
faceted by .schema If .schema is not specified, facet across all schema
in fndata.}

\item{.lnsize}{A vector c(low, high) of the range of edge (line) widths to plot. Set smaller
values (e.g., c(1, 3)) for plots with many facets to avoid overlapping edges.}
}
\value{
A network plot of similarity between groups.
}
\description{
Create a network diagram of similarity between groups from a long data frame
   of diversity measures.
}
\details{
NOTE: THIS FUNCTION PLOTS SIMILARITY (1 - DIVERSITY).

   This function returns a network diagram of similarity among all unique groups
   found in the 'group_from' and 'group_to' columns. If more than one measure
   or schema are specified, the returned plot is faceted. If .measure or .schema
   are not specified, the returned plot is faceted across all measures and schema
   in fndata. fndata must be in long format, with columns:  'schema', 'measure',
   'value', 'group_from`, `group_to`.

   Why plot similarity and not diversity?

   A network plot demonstrates magnitude through colour and edge (line) width.
   The natural interpretation of colour intensity and edge width is of a measure
   of the similarity between nodes (i.e., where an increased value corresponds
   to increased similarity). Plotting diversity necessitates inverting each
   scale, and makes interpretation more difficult.
}
