% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkmatrix.R
\name{mkmatrix}
\alias{mkmatrix}
\title{Create an example ebdmc matrix}
\usage{
mkmatrix(.nrow = 5, .ncol = 10, .type = c("cim", "im", "cdm"), .knit = TRUE)
}
\arguments{
\item{.nrow}{integer: number of rows to generate}

\item{.ncol}{integer: number of columns to generate}

\item{.type}{string: one of: 'cim', 'im', or 'cdm'}

\item{.knit}{boolean: use knitr::kable() to style the output}
}
\value{
an example matrix
}
\description{
Create an example matrix for documentation in the ebdmc project.
}
