% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msafely.R
\name{msafely}
\alias{msafely}
\title{Replace unsafe names of measures of beta diversity}
\usage{
msafely(.measures)
}
\arguments{
\item{.measure}{a string or vector: the unsafe name of the measure(s).}
}
\value{
a string or vector: the safe name of the measure(s), where applicable.
}
\description{
Replace unsafe (e.g., for column) names of measures of beta diversity
  in betadiver() with safe names.
}
