% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_awc.R
\name{neighbour_awc}
\alias{neighbour_awc}
\alias{nawc_shannon}
\title{Calculate the neighbour-adjusted Wallace coefficient}
\usage{
neighbour_awc(clusters, i)

nawc_shannon(clusters)
}
\arguments{
\item{clusters_df}{A set of cluster identities (cluster memberships) for each member in a
community.}
}
\value{
The diversity index.
}
\description{
Calculate the neighbour-adjusted Wallace coefficient (nAWC).
}
\details{
From dorbarker's nAWC \url{https://github.com/dorbarker/nAWC}.
}
