% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_fns.R
\name{partitioncoefficient}
\alias{partitioncoefficient}
\alias{rand}
\alias{wallace}
\title{Calculate a partition coefficient}
\usage{
rand(a, b, c, d)

wallace(a, b, c)
}
\arguments{
\item{a, b, c, d}{The parameters of a 'mismatch' matrix}
}
\value{
The partition coefficient.
}
\description{
Calculate a partition coefficient from a 'mismatch' matrix; the agreement/
  disagreement between two clustering schemes.
}
\details{
From dorbarker's nAWC \url{https://github.com/dorbarker/nAWC}.
}
