% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beta.R
\name{print_beta}
\alias{print_beta}
\title{Print the name of a measure of beta diversity in-text using a beta symbol}
\usage{
print_beta(.measure)
}
\arguments{
\item{.measure}{string: the name of the measure.}
}
\value{
a string: .measure in the format b_.measure (beta subscript .measure).
}
\description{
Print the name of a measure of beta diversity using a beta symbol in-text
  (e.g., for RMD/QMD documents), in the format beta subscript measure.
}
